fix redux
