+++ src/redux/actions/shopActions.ts
@@ -1,5 +1,6 @@
 import ShopTypes from "@redux/types/shopTypes";
 import { IEmployee } from "src/types/employee";
+import { IService } from "src/types/service";
 
 interface IShopActionBase {
   type: ShopTypes;
@@ -13,9 +14,18 @@ export interface IShopActionSetShopEmployees extends IShopActionBase {
   payload: { employees: IEmployee[] | null };
 }
 
+export interface IServiceActionReadShopServices extends IShopActionBase {
+  payload: void;
+}
+
+export interface IServiceActionSetShopServices extends IShopActionBase {
+  payload: { services: IService[] };
+}
+
 export type IShopAction =
   | IShopActionReadShopEmployees
-  | IShopActionSetShopEmployees;
+  | IShopActionSetShopEmployees
+  | IServiceActionSetShopServices;
 
 export const readShopEmployees = (payload: void) => ({
   type: ShopTypes.READ_SHOP_EMPLOYEES,
@@ -26,5 +36,17 @@ export const setShopEmployees = (
   payload: IShopActionSetShopEmployees["payload"]
 ) => ({
   type: ShopTypes.SET_SHOP_EMPLOYEES,
+  payload,
+});
+
+export const readShopServices = (payload: void) => ({
+  type: ShopTypes.READ_SHOP_SERVICES,
+  payload,
+});
+
+export const setShopServices = (
+  payload: IServiceActionSetShopServices["payload"]
+) => ({
+  type: ShopTypes.SET_SHOP_SERVICES,
   payload,
 });
