+++ src/redux/types/serviceTypes.ts
@@ -1,7 +1,5 @@
 enum ServiceTypes {
   SET_SERVICE = "SET SERVICE",
-  SET_SERVICES = "SET SERVICES",
-  READ_SERVICES = "READ SERVICES",
   CREATE_SERVICE = "CREATE SERVICE",
   DELETE_SERVICE = "DELETE SERVICE",
 }
