+++ src/redux/types/shopTypes.ts
@@ -1,6 +1,8 @@
 enum ShopTypes {
   READ_SHOP_EMPLOYEES = "READ EMPLOYEES",
   SET_SHOP_EMPLOYEES = "SET EMPLOYEES",
+  READ_SHOP_SERVICES = "READ SERVICES",
+  SET_SHOP_SERVICES = "SET SERVICES",
 }
 
 export default ShopTypes;
